no complete
I update some things. I declared a char to the variable answer, I added
some comments to know what does each part of the code and I added a new
do-while loop to do the whole thing again if the user wants. The only
detail is that the factorial only can be of numbers between 1 to 12.
I’ll update this code later with a “BigIntegerLibrary.hh”